[WIP] Moving enums
