+++ src/Entities/Method/Get.php
@@ -2,8 +2,6 @@
 
 namespace Somecode\OpenApi\Entities\Method;
 
-use Somecode\OpenApi\Enums\RequestMethod;
-
 class Get extends Method
 {
     public function method(): RequestMethod
