+++ src/Entities/Method/Post.php
@@ -2,8 +2,6 @@
 
 namespace Somecode\OpenApi\Entities\Method;
 
-use Somecode\OpenApi\Enums\RequestMethod;
-
 class Post extends Method
 {
     public function method(): RequestMethod
