+++ src/Entities/Method/Put.php
@@ -2,8 +2,6 @@
 
 namespace Somecode\OpenApi\Entities\Method;
 
-use Somecode\OpenApi\Enums\RequestMethod;
-
 class Put extends Method
 {
     public function method(): RequestMethod
